workers/src/workers/pdf_worker/pdf_worker.py
"""
PDF Worker implementation
Handles PDF parsing, layout analysis, figure/table detection, and chunking
"""

import asyncio
import logging
from typing import Dict, List, Any
import fitz  # PyMuPDF
import pdfplumber
from PIL import Image
import io
import structlog

logger = structlog.get_logger(__name__)


class PDFWorker:
    """PDF processing worker for document ingestion"""
    
    def __init__(self):
        self.logger = logger.bind(worker="pdf_worker")
    
    async def process_document(self, document_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Process a PDF document for ingestion
        
        Args:
            document_data: Document metadata and S3 location
            
        Returns:
            Processing results with chunks, figures, and tables
        """
        try:
            self.logger.info("Starting PDF processing", document_id=document_data.get('document_id'))
            
            # TODO: Download PDF from S3
            # pdf_path = await self.download_from_s3(document_data['s3_key'])
            
            # For now, use a placeholder
            pdf_path = "/tmp/sample.pdf"
            
            # Parse PDF structure
            structure = await self.parse_pdf_structure(pdf_path)
            
            # Detect figures and tables
            figures = await self.detect_figures(pdf_path, structure)
            tables = await self.extract_tables(pdf_path, structure)
            
            # Create text chunks
            chunks = await self.create_chunks(structure)
            
            result = {
                'document_id': document_data.get('document_id'),
                'chunks': chunks,
                'figures': figures,
                'tables': tables,
                'metadata': {
                    'total_pages': structure['total_pages'],
                    'sections': structure['sections'],
                    'processing_time': 0.0  # TODO: Add timing
                }
            }
            
            self.logger.info("PDF processing completed", 
                           document_id=document_data.get('document_id'),
                           chunks_count=len(chunks),
                           figures_count=len(figures),
                           tables_count=len(tables))
            
            return result
            
        except Exception as e:
            self.logger.error("PDF processing failed", 
                            document_id=document_data.get('document_id'),
                            error=str(e))
            raise
    
    async def parse_pdf_structure(self, pdf_path: str) -> Dict[str, Any]:
        """Parse PDF structure and extract sections"""
        try:
            doc = fitz.open(pdf_path)
            structure = {
                'total_pages': len(doc),
                'sections': [],
                'text_by_page': []
            }
            
            for page_num in range(len(doc)):
                page = doc[page_num]
                text = page.get_text()
                structure['text_by_page'].append({
                    'page': page_num + 1,
                    'text': text,
                    'bbox': page.rect
                })
            
            # TODO: Implement section detection (abstract, introduction, methods, etc.)
            structure['sections'] = self._detect_sections(structure['text_by_page'])
            
            doc.close()
            return structure
            
        except Exception as e:
            self.logger.error("PDF structure parsing failed", error=str(e))
            raise
    
    async def detect_figures(self, pdf_path: str, structure: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Detect figures in the PDF"""
        figures = []
        
        try:
            doc = fitz.open(pdf_path)
            
            for page_num in range(len(doc)):
                page = doc[page_num]
                
                # Look for image blocks
                image_blocks = page.get_image_info()
                
                for img in image_blocks:
                    figure = {
                        'page': page_num + 1,
                        'bbox': img['bbox'],
                        'figure_number': self._extract_figure_number(page, img['bbox']),
                        'caption': self._extract_caption(page, img['bbox']),
                        's3_key': None  # TODO: Extract and upload figure image
                    }
                    figures.append(figure)
            
            doc.close()
            
        except Exception as e:
            self.logger.error("Figure detection failed", error=str(e))
        
        return figures
    
    async def extract_tables(self, pdf_path: str, structure: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Extract tables from the PDF"""
        tables = []
        
        try:
            with pdfplumber.open(pdf_path) as pdf:
                for page_num, page in enumerate(pdf.pages):
                    page_tables = page.extract_tables()
                    
                    for table_idx, table in enumerate(page_tables):
                        if table and len(table) > 1:  # Valid table
                            table_data = {
                                'page': page_num + 1,
                                'table_number': table_idx + 1,
                                'data': table,
                                'bbox': page.bbox,
                                'title': self._extract_table_title(page, table_idx),
                                'caption': self._extract_table_caption(page, table_idx)
                            }
                            tables.append(table_data)
            
        except Exception as e:
            self.logger.error("Table extraction failed", error=str(e))
        
        return tables
    
    async def create_chunks(self, structure: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Create text chunks for embedding"""
        chunks = []
        
        try:
            for section in structure['sections']:
                # Create chunks within each section
                section_chunks = self._chunk_text(section['text'], 
                                                section['name'], 
                                                section['page_from'], 
                                                section['page_to'])
                chunks.extend(section_chunks)
            
        except Exception as e:
            self.logger.error("Chunk creation failed", error=str(e))
        
        return chunks
    
    def _detect_sections(self, text_by_page: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """Detect document sections"""
        # TODO: Implement section detection logic
        # This is a placeholder implementation
        sections = [
            {
                'name': 'abstract',
                'page_from': 1,
                'page_to': 1,
                'text': text_by_page[0]['text'][:1000] if text_by_page else ''
            }
        ]
        return sections
    
    def _chunk_text(self, text: str, section: str, page_from: int, page_to: int) -> List[Dict[str, Any]]:
        """Split text into chunks"""
        # Simple chunking by sentences/paragraphs
        # TODO: Implement more sophisticated chunking
        chunk_size = 800  # target tokens
        chunks = []
        
        # Split by paragraphs first
        paragraphs = text.split('\n\n')
        current_chunk = ""
        current_pages = []
        
        for i, para in enumerate(paragraphs):
            if len(current_chunk) + len(para) > chunk_size and current_chunk:
                chunks.append({
                    'text': current_chunk.strip(),
                    'section': section,
                    'page_from': min(current_pages) if current_pages else page_from,
                    'page_to': max(current_pages) if current_pages else page_to,
                    'metadata': {
                        'chunk_type': 'text',
                        'word_count': len(current_chunk.split())
                    }
                })
                current_chunk = para
                current_pages = [page_from]  # Simplified
            else:
                current_chunk += "\n\n" + para if current_chunk else para
                current_pages.append(page_from)  # Simplified
        
        # Add final chunk
        if current_chunk:
            chunks.append({
                'text': current_chunk.strip(),
                'section': section,
                'page_from': min(current_pages) if current_pages else page_from,
                'page_to': max(current_pages) if current_pages else page_to,
                'metadata': {
                    'chunk_type': 'text',
                    'word_count': len(current_chunk.split())
                }
            })
        
        return chunks
    
    def _extract_figure_number(self, page, bbox) -> str:
        """Extract figure number from page"""
        # TODO: Implement figure number extraction
        return "Figure 1"
    
    def _extract_caption(self, page, bbox) -> str:
        """Extract figure caption from page"""
        # TODO: Implement caption extraction
        return "Sample figure caption"
    
    def _extract_table_title(self, page, table_idx) -> str:
        """Extract table title"""
        # TODO: Implement table title extraction
        return f"Table {table_idx + 1}"
    
    def _extract_table_caption(self, page, table_idx) -> str:
        """Extract table caption"""
        # TODO: Implement table caption extraction
        return "Sample table caption"
