workers/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0

# PDF processing
pypdf2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
pytesseract==0.3.10
pillow==10.1.0

# Data processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
xlrd==2.0.1

# ML/AI
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Vector database
pgvector==0.2.4

# Plotting & visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Messaging
nats-py==2.6.0
redis==5.0.1

# AWS
boto3==1.34.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3
structlog==23.2.0
